requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# LLM and RAG Dependencies
openai==1.3.7
langchain==0.0.340
langchain-openai==0.0.2
tiktoken==0.5.2
sentence-transformers==2.2.2

# Vector Database
faiss-cpu==1.7.4
numpy==1.24.3
scipy==1.11.4

# Caching and Database
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# HTTP and Networking
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data Processing
pandas==2.1.4
pyarrow==14.0.1
polars==0.19.19

# Compliance and Security
cryptography==41.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0
loguru==0.7.2

# Explainability
lime==0.2.0.1
shap==0.43.0
scikit-learn==1.3.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Azure SDK (for deployment)
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.19.0
azure-monitor-opentelemetry==1.1.1

# Performance
asyncio==3.4.3
aiofiles==23.2.1
orjson==3.9.10

# Configuration Management
python-dotenv==1.0.0
pyyaml==6.0.1
toml==0.10.2

# Data Validation
marshmallow==3.20.1
cerberus==1.3.5

# Utilities
click==8.1.7
rich==13.7.0
tqdm==4.66.1